+++ js/main.js
@@ -2,12 +2,14 @@ import { openBigPhotoPopup } from './big-picture-popup.js';
 import { getData } from './api.js';
 import { closeUploadOverlay, setUploadFormSubmit } from './upload-form.js';
 import { renderPictures, showDataError } from './render-pictures.js';
+import { setFilter } from './filters.js';
 
 let pictures;
 
 getData().then((photos) => {
   pictures = photos;
   renderPictures(pictures);
+  setFilter(pictures);
 })
   .catch(showDataError);
